linting and test fixes
