+++ services/QuillLMS/app/assets/stylesheets/pages/rosters.scss
@@ -1,5 +1,5 @@
 .roster-input-container {
-  padding-top: 20px;
+  display: flex;
 }
 
 .roster-school-id {
