+++ services/QuillLMS/config/routes.rb
@@ -513,7 +513,9 @@ EmpiricalGrammar::Application.routes.draw do
     get '/concepts/concepts_in_use', to: 'concepts#concepts_in_use', only: [:csv], defaults: { format: 'csv' }
     resources :concepts
     resources :evidence, only: [:index]
-    resources :rosters, only: [:index]
+    resources :rosters, only: [:index] do
+      post :upload_teachers_and_students, on: :collection
+    end
     resources :standard_levels, only: [:index, :create, :update]
     resources :standards, only: [:index, :create, :update]
     resources :content_partners, only: [:index, :create, :update]
@@ -554,7 +556,6 @@ EmpiricalGrammar::Application.routes.draw do
       collection do
         post :search
         get :search, to: 'schools#index'
-        post :upload_teachers_and_students
       end
       member do
         get :edit_subscription
